Register a global hotkey
Build DDHotKey as a static library so we can compile it with ARC turned
on and our project with ARC turned off. Since DDHotKey is written with
ARC, if I include its headers directly, I'd need to enable ARC on my
project in order to compile it.

Initialise an `NSApplication` to give us an environment in which to
register the global hotkey.

A new class `Invert` will respond to the hotkey and respond accordingly.